% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvalueOptimizer.R
\docType{class}
\name{CvalueOptimizer-class}
\alias{CvalueOptimizer-class}
\alias{CvalueOptimizer}
\title{CvalueOptimizer class}
\description{
Virtual class for Random forest. 
Contains all fields and methods used in all Forest subclasses.
}

