% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvalueOptimizerClassification.R
\docType{class}
\name{CvalueOptimizerClassification-class}
\alias{CvalueOptimizerClassification-class}
\alias{CvalueOptimizerClassification}
\title{Classification forest class}
\description{
Subclass for classification forest. 
Contains all fields and methods used special for classification forests.
}

