% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvalueOptimizerRegression.R
\docType{class}
\name{CvalueOptimizerRegression-class}
\alias{CvalueOptimizerRegression-class}
\alias{CvalueOptimizerRegression}
\title{Classification forest class}
\description{
Subclass for classification forest. 
Contains all fields and methods used special for classification forests.
}

