% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvalueOptimizerSurvival.R
\docType{class}
\name{CvalueOptimizerSurvival-class}
\alias{CvalueOptimizerSurvival-class}
\alias{CvalueOptimizerSurvival}
\title{Classification forest class}
\description{
Subclass for classification forest. 
Contains all fields and methods used special for classification forests.
}

