% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.blockForest}
\alias{importance.blockForest}
\alias{importance}
\title{blockForest variable importance}
\usage{
\method{importance}{blockForest}(x, ...)
}
\arguments{
\item{x}{blockForest object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Variable importance measures.
}
\description{
Extract variable importance of blockForest object.
}
\seealso{
\code{\link{blockForest}}
}
\author{
Marvin N. Wright
}
