% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions.blockForest.prediction}
\alias{predictions.blockForest.prediction}
\alias{predictions}
\title{blockForest predictions}
\usage{
\method{predictions}{blockForest.prediction}(x, ...)
}
\arguments{
\item{x}{blockForest prediction object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Predictions: Classes for Classification forests, Numerical values for Regressions forests and the estimated survival functions for all individuals for Survival forests.
}
\description{
Extract predictions of blockForest prediction object.
}
\seealso{
\code{\link{blockForest}}
}
\author{
Marvin N. Wright
}
