% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions.blockForest}
\alias{predictions.blockForest}
\title{blockForest predictions}
\usage{
\method{predictions}{blockForest}(x, ...)
}
\arguments{
\item{x}{blockForest object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Predictions: Classes for Classification forests, Numerical values for Regressions forests and the estimated survival functions for all individuals for Survival forests.
}
\description{
Extract training data predictions of blockForest object.
}
\seealso{
\code{\link{blockForest}}
}
\author{
Marvin N. Wright
}
