% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.blockForest.prediction}
\alias{print.blockForest.prediction}
\title{Print blockForest prediction}
\usage{
\method{print}{blockForest.prediction}(x, ...)
}
\arguments{
\item{x}{Object of class 'blockForest.prediction'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of blockForest prediction object.
}
\author{
Marvin N. Wright
}
