% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{timepoints.blockForest.prediction}
\alias{timepoints.blockForest.prediction}
\title{blockForest timepoints}
\usage{
\method{timepoints}{blockForest.prediction}(x, ...)
}
\arguments{
\item{x}{blockForest Survival prediction object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Unique death times
}
\description{
Extract unique death times of blockForest Survival prediction object.
}
\seealso{
\code{\link{blockForest}}
}
\author{
Marvin N. Wright
}
