% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{timepoints.blockForest}
\alias{timepoints.blockForest}
\alias{timepoints}
\title{blockForest timepoints}
\usage{
\method{timepoints}{blockForest}(x, ...)
}
\arguments{
\item{x}{blockForest Survival forest object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Unique death times
}
\description{
Extract unique death times of blockForest Survival forest
}
\seealso{
\code{\link{blockForest}}
}
\author{
Marvin N. Wright
}
