% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockfor.R
\name{blockfor}
\alias{blockfor}
\title{blockfor}
\usage{
blockfor(X, y, num_trees = 50, mtrys = NULL, min_node_size = NULL,
  replace = TRUE, probability = FALSE, splitrule = NULL, block,
  num_treesorder = 500, nsets = 100, num_treesoptim = 100)
}
\arguments{
\item{X}{matrix. covariate matrix. observations in rows, variables in columns.}

\item{y}{target variable. See 'Example.R'.}

\item{num_trees}{Number of trees in block forest.}

\item{mtrys}{Vector of length equal to the number of blocks. For each block: number of variables to possibly split at in each node. Leave unspecified.}

\item{min_node_size}{Minimal node size. Default 1 for classification, 5 for regression and 3 for survival.}

\item{replace}{Sample with replacement. Default TRUE.}

\item{probability}{Grow a probability forest. Default FALSE.}

\item{splitrule}{Splitrule to use in trees. Default "Gini" for classification forests, "Variance" for regression and probability forests and "Logrank" for survival forests.}

\item{block}{list. A list of length equal to the number M of blocks considered. Each
entry contains the vector of column indices in 'X' of the covariates in one of the M blocks.}

\item{num_treesorder}{B_order in Supplementary Data}

\item{nsets}{N_sets in Supplementary Data}

\item{num_treesoptim}{B_optim in Supplementary Data}
}
\description{
Implements Random Forests (Breiman 2001) with emphasis on simplicity. 
Uses reference classes and only plain \code{R}. 
Not optimized for computation speed. 
Allows rapid prototyping of RF-type algorithms.
}
\examples{
\donttest{
library(blockForest)

# Classification
blockfor(Species ~ ., iris)

# Prediction
train_idx <- sample(nrow(iris), 2/3 * nrow(iris))
iris_train <- iris[train_idx, ]
iris_test <- iris[-train_idx, ]
rf_iris <- blockfor(Species ~ ., data = iris_train)
pred_iris <- rf_iris$predict(iris_test)
table(iris_test$Species, pred_iris)
}

}
\references{
Breiman, L. (2001). Random forests. Mach Learn, 45(1), 5-32. \cr
}
\author{
Marvin N. Wright
}
